+++ packages/ai-constructs/src/conversation/runtime/conversation_turn_response_sender.test.ts
@@ -3,8 +3,9 @@ import assert from 'node:assert';
 import {
   ConversationTurnResponseSender,
   MutationResponseInput,
+  MutationStreamingResponseInput,
 } from './conversation_turn_response_sender';
-import { ConversationTurnEvent } from './types';
+import { ConversationTurnEvent, StreamingResponseChunk } from './types';
 import { ContentBlock } from '@aws-sdk/client-bedrock-runtime';
 import {
   GraphqlRequest,
@@ -129,6 +130,108 @@ void describe('Conversation turn response sender', () => {
             },
           ],
           associatedUserMessageId: event.currentMessageId,
+        },
+      },
+    });
+  });
+
+  void it('sends streaming response chunk back to appsync', async () => {
+    const graphqlRequestExecutor = new GraphqlRequestExecutor('', '', '');
+    const executeGraphqlMock = mock.method(
+      graphqlRequestExecutor,
+      'executeGraphql',
+      () =>
+        // Mock successful Appsync response
+        Promise.resolve()
+    );
+    const sender = new ConversationTurnResponseSender(
+      event,
+      graphqlRequestExecutor
+    );
+    const chunk: StreamingResponseChunk = {
+      accumulatedTurnContent: [{ text: 'testAccumulatedMessageContent' }],
+      associatedUserMessageId: 'testAssociatedUserMessageId',
+      contentBlockIndex: 1,
+      contentBlockDeltaIndex: 2,
+      conversationId: 'testConversationId',
+      contentBlockText: 'testBlockText',
+    };
+    await sender.sendResponseChunk(chunk);
+
+    assert.strictEqual(executeGraphqlMock.mock.calls.length, 1);
+    const request = executeGraphqlMock.mock.calls[0]
+      .arguments[0] as GraphqlRequest<MutationStreamingResponseInput>;
+    assert.deepStrictEqual(request, {
+      query:
+        '\n' +
+        '        mutation PublishModelResponse($input: testResponseMutationInputTypeName!) {\n' +
+        '            testResponseMutationName(input: $input) {\n' +
+        '                testSelectionSet\n' +
+        '            }\n' +
+        '        }\n' +
+        '    ',
+      variables: {
+        input: chunk,
+      },
+    });
+  });
+
+  void it('serializes tool use input to JSON when streaming', async () => {
+    const graphqlRequestExecutor = new GraphqlRequestExecutor('', '', '');
+    const executeGraphqlMock = mock.method(
+      graphqlRequestExecutor,
+      'executeGraphql',
+      () =>
+        // Mock successful Appsync response
+        Promise.resolve()
+    );
+    const sender = new ConversationTurnResponseSender(
+      event,
+      graphqlRequestExecutor
+    );
+    const toolUseBlock: ContentBlock.ToolUseMember = {
+      toolUse: {
+        name: 'testTool',
+        toolUseId: 'testToolUseId',
+        input: {
+          testPropertyKey: 'testPropertyValue',
+        },
+      },
+    };
+    const chunk: StreamingResponseChunk = {
+      accumulatedTurnContent: [toolUseBlock],
+      associatedUserMessageId: 'testAssociatedUserMessageId',
+      contentBlockIndex: 1,
+      contentBlockDeltaIndex: 2,
+      conversationId: 'testConversationId',
+      contentBlockText: 'testBlockText',
+    };
+    await sender.sendResponseChunk(chunk);
+
+    assert.strictEqual(executeGraphqlMock.mock.calls.length, 1);
+    const request = executeGraphqlMock.mock.calls[0]
+      .arguments[0] as GraphqlRequest<MutationStreamingResponseInput>;
+    assert.deepStrictEqual(request, {
+      query:
+        '\n' +
+        '        mutation PublishModelResponse($input: testResponseMutationInputTypeName!) {\n' +
+        '            testResponseMutationName(input: $input) {\n' +
+        '                testSelectionSet\n' +
+        '            }\n' +
+        '        }\n' +
+        '    ',
+      variables: {
+        input: {
+          ...chunk,
+          accumulatedTurnContent: [
+            {
+              toolUse: {
+                input: JSON.stringify(toolUseBlock.toolUse.input),
+                name: toolUseBlock.toolUse.name,
+                toolUseId: toolUseBlock.toolUse.toolUseId,
+              },
+            },
+          ],
         },
       },
     });