+++ packages/integration-tests/src/test-e2e/deployment/access_testing_project.deployment.test.ts
@@ -1,0 +1,4 @@
+import { AccessTestingProjectTestProjectCreator } from '../../test-project-setup/access_testing_project.js';
+import { defineDeploymentTest } from './deployment.test.template.js';
+
+defineDeploymentTest(new AccessTestingProjectTestProjectCreator());